Create IronA11yKeysMixin
This creates a mixin that implements key bindings for non-legacy Polymer
2.0 elements.  The mixin does not look at the keyBindings property of
behaviors, all keyBindings must be declared as a static property of a
class in the hierarchy.

A new test has been created to check the mixin based key bindings.  All
tests passed on Chrome 57 and Firefox 52 under Fedora 25.  travis-ci
will likely fail due to use of variants.